requirements.txt
# --- Core Backend ---
fastapi
uvicorn
python-dotenv
orjson

# --- LangChain & Integrations ---
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai

# --- Embeddings & Vector Store ---
sentence-transformers
faiss-cpu

# --- Database / Cache ---
diskcache


# --- File & Data Handling ---
pypdf
numpy